+++ test/integration/integration_test.go
@@ -17,10 +17,8 @@ limitations under the License.
 package integration
 
 import (
-	"errors"
 	"flag"
 	"os"
-	"strings"
 	"testing"
 
 	"github.com/ExpediaGroup/container-startup-autoscaler/internal/pod"
@@ -609,7 +607,7 @@ func TestDeploymentScaleWhenUnknownResourcesAll(t *testing.T) {
 
 	assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
 
-	ensureEvents(
+	assertEvents(
 		t,
 		csaEventReasonScaling,
 		[]string{
@@ -663,7 +661,7 @@ func TestDeploymentScaleWhenUnknownResourcesCpu(t *testing.T) {
 
 	assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
 
-	ensureEvents(
+	assertEvents(
 		t,
 		csaEventReasonScaling,
 		[]string{
@@ -717,7 +715,7 @@ func TestDeploymentScaleWhenUnknownResourcesMemory(t *testing.T) {
 
 	assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
 
-	ensureEvents(
+	assertEvents(
 		t,
 		csaEventReasonScaling,
 		[]string{
@@ -902,7 +900,7 @@ func TestValidationFailure(t *testing.T) {
 		require.Empty(t, statusAnn.Scale.LastFailed)
 	}
 
-	ensureEvents(t, csaEventReasonValidation, []string{csaStatusMessageValidationError}, namespace, names)
+	assertEvents(t, csaEventReasonValidation, []string{csaStatusMessageValidationError}, namespace, names)
 }
 
 // Helpers -------------------------------------------------------------------------------------------------------------
@@ -963,7 +961,7 @@ func testWorkflow(
 	}
 	assertPostStartupEnactedRestartFunc(t, annotations, podStatusAnn)
 
-	ensureEvents(
+	assertEvents(
 		t,
 		csaEventReasonScaling,
 		[]string{
@@ -975,226 +973,10 @@ func testWorkflow(
 	)
 }
 
-func assertStartupEnacted(
-	t *testing.T,
-	annotations csaQuantityAnnotations,
-	podStatusAnn map[*v1.Pod]pod.StatusAnnotation,
-	expectStartupProbe bool,
-	expectReadinessProbe bool,
-	expectStatusCommandedEnactedEmpty bool,
-) {
-	if (!expectStartupProbe && !expectReadinessProbe) || (expectStartupProbe && expectReadinessProbe) {
-		panic(errors.New("only one of expectStartupProbe/expectReadinessProbe must be true"))
-	}
-
-	for kubePod, statusAnn := range podStatusAnn {
-		for _, c := range kubePod.Spec.Containers {
-			var expectCpuR, expectCpuL, expectMemoryR, expectMemoryL string
-
-			if c.Name == echoServerName {
-				expectCpuR, expectCpuL = annotations.CpuStartupRequestsLimits()
-				expectMemoryR, expectMemoryL = annotations.MemoryStartupRequestsLimits()
-			} else if c.Name == echoServerNonTargetContainerName {
-				expectCpuR, expectCpuL = echoServerNonTargetContainerCpuRequests, echoServerNonTargetContainerCpuLimits
-				expectMemoryR, expectMemoryL = echoServerNonTargetContainerMemoryRequests, echoServerNonTargetContainerMemoryLimits
-			} else {
-				panic(errors.New("container name unrecognized"))
-			}
-
-			if expectStartupProbe {
-				require.NotNil(t, c.StartupProbe)
-			} else {
-				require.Nil(t, c.StartupProbe)
-			}
-			if expectReadinessProbe {
-				require.NotNil(t, c.ReadinessProbe)
-			} else {
-				require.Nil(t, c.ReadinessProbe)
-			}
-			cpuR := c.Resources.Requests[v1.ResourceCPU]
-			require.Equal(t, expectCpuR, cpuR.String())
-			cpuL := c.Resources.Limits[v1.ResourceCPU]
-			require.Equal(t, expectCpuL, cpuL.String())
-			memoryR := c.Resources.Requests[v1.ResourceMemory]
-			require.Equal(t, expectMemoryR, memoryR.String())
-			memoryL := c.Resources.Limits[v1.ResourceMemory]
-			require.Equal(t, expectMemoryL, memoryL.String())
-		}
-
-		for _, s := range kubePod.Status.ContainerStatuses {
-			var expectCpuR, expectCpuL, expectMemoryR, expectMemoryL string
-
-			if s.Name == echoServerName {
-				expectCpuR, expectCpuL = annotations.CpuStartupRequestsLimits()
-				expectMemoryR, expectMemoryL = annotations.MemoryStartupRequestsLimits()
-
-				// See comment in targetcontaineraction.go
-				if expectStartupProbe {
-					require.False(t, *s.Started)
-				} else {
-					require.True(t, *s.Started)
-				}
-				require.False(t, s.Ready)
-			} else if s.Name == echoServerNonTargetContainerName {
-				expectCpuR, expectCpuL = echoServerNonTargetContainerCpuRequests, echoServerNonTargetContainerCpuLimits
-				expectMemoryR, expectMemoryL = echoServerNonTargetContainerMemoryRequests, echoServerNonTargetContainerMemoryLimits
-			} else {
-				panic(errors.New("container name unrecognized"))
-			}
-
-			require.NotNil(t, s.State.Running)
-			cpuR := s.Resources.Requests[v1.ResourceCPU]
-			require.Equal(t, expectCpuR, cpuR.String())
-			cpuL := s.Resources.Limits[v1.ResourceCPU]
-			require.Equal(t, expectCpuL, cpuL.String())
-			memoryR := s.Resources.Requests[v1.ResourceMemory]
-			require.Equal(t, expectMemoryR, memoryR.String())
-			memoryL := s.Resources.Limits[v1.ResourceMemory]
-			require.Equal(t, expectMemoryL, memoryL.String())
-		}
-
-		require.Equal(t, csaStatusMessageStartupEnacted, statusAnn.Status)
-		require.NotEmpty(t, statusAnn.LastUpdated)
-
-		require.Equal(t, expectStartupProbe, statusAnn.States.StartupProbe.Bool())
-		require.Equal(t, expectReadinessProbe, statusAnn.States.ReadinessProbe.Bool())
-		require.Equal(t, podcommon.StateContainerRunning, statusAnn.States.Container)
-		if expectStartupProbe {
-			require.Equal(t, podcommon.StateBoolFalse, statusAnn.States.Started)
-		} else {
-			require.Equal(t, podcommon.StateBoolTrue, statusAnn.States.Started)
-		}
-		require.Equal(t, podcommon.StateBoolFalse, statusAnn.States.Ready)
-		require.Equal(t, podcommon.StateResourcesStartup, statusAnn.States.Resources)
-		require.Equal(t, podcommon.StateStatusResourcesContainerResourcesMatch, statusAnn.States.StatusResources)
-
-		if expectStatusCommandedEnactedEmpty {
-			require.Empty(t, statusAnn.Scale.LastCommanded)
-			require.Empty(t, statusAnn.Scale.LastEnacted)
-		} else {
-			require.NotEmpty(t, statusAnn.Scale.LastCommanded)
-			require.NotEmpty(t, statusAnn.Scale.LastEnacted)
-		}
-		require.Empty(t, statusAnn.Scale.LastFailed)
-	}
-}
-
-func assertPostStartupEnacted(
-	t *testing.T,
-	annotations csaQuantityAnnotations,
-	podStatusAnn map[*v1.Pod]pod.StatusAnnotation,
-	expectStartupProbe bool,
-	expectReadinessProbe bool,
-) {
-	for kubePod, statusAnn := range podStatusAnn {
-		for _, c := range kubePod.Spec.Containers {
-			var expectCpuR, expectCpuL, expectMemoryR, expectMemoryL string
-
-			if c.Name == echoServerName {
-				expectCpuR, expectCpuL = annotations.CpuPostStartupRequestsLimits()
-				expectMemoryR, expectMemoryL = annotations.MemoryPostStartupRequestsLimits()
-			} else if c.Name == echoServerNonTargetContainerName {
-				expectCpuR, expectCpuL = echoServerNonTargetContainerCpuRequests, echoServerNonTargetContainerCpuLimits
-				expectMemoryR, expectMemoryL = echoServerNonTargetContainerMemoryRequests, echoServerNonTargetContainerMemoryLimits
-			} else {
-				panic(errors.New("container name unrecognized"))
-			}
-
-			if expectStartupProbe {
-				require.NotNil(t, c.StartupProbe)
-			} else {
-				require.Nil(t, c.StartupProbe)
-			}
-			if expectReadinessProbe {
-				require.NotNil(t, c.ReadinessProbe)
-			} else {
-				require.Nil(t, c.ReadinessProbe)
-			}
-			cpuR := c.Resources.Requests[v1.ResourceCPU]
-			require.Equal(t, expectCpuR, cpuR.String())
-			cpuL := c.Resources.Limits[v1.ResourceCPU]
-			require.Equal(t, expectCpuL, cpuL.String())
-			memoryR := c.Resources.Requests[v1.ResourceMemory]
-			require.Equal(t, expectMemoryR, memoryR.String())
-			memoryL := c.Resources.Limits[v1.ResourceMemory]
-			require.Equal(t, expectMemoryL, memoryL.String())
-		}
-
-		for _, s := range kubePod.Status.ContainerStatuses {
-			var expectCpuR, expectCpuL, expectMemoryR, expectMemoryL string
-
-			if s.Name == echoServerName {
-				expectCpuR, expectCpuL = annotations.CpuPostStartupRequestsLimits()
-				expectMemoryR, expectMemoryL = annotations.MemoryPostStartupRequestsLimits()
-
-				// See comment in targetcontaineraction.go
-				require.True(t, *s.Started)
-				require.True(t, s.Ready)
-			} else if s.Name == echoServerNonTargetContainerName {
-				expectCpuR, expectCpuL = echoServerNonTargetContainerCpuRequests, echoServerNonTargetContainerCpuLimits
-				expectMemoryR, expectMemoryL = echoServerNonTargetContainerMemoryRequests, echoServerNonTargetContainerMemoryLimits
-			} else {
-				panic(errors.New("container name unrecognized"))
-			}
-
-			require.NotNil(t, s.State.Running)
-			cpuR := s.Resources.Requests[v1.ResourceCPU]
-			require.Equal(t, expectCpuR, cpuR.String())
-			cpuL := s.Resources.Limits[v1.ResourceCPU]
-			require.Equal(t, expectCpuL, cpuL.String())
-			memoryR := s.Resources.Requests[v1.ResourceMemory]
-			require.Equal(t, expectMemoryR, memoryR.String())
-			memoryL := s.Resources.Limits[v1.ResourceMemory]
-			require.Equal(t, expectMemoryL, memoryL.String())
-		}
-
-		require.Equal(t, csaStatusMessagePostStartupEnacted, statusAnn.Status)
-		require.NotEmpty(t, statusAnn.LastUpdated)
-
-		require.Equal(t, expectStartupProbe, statusAnn.States.StartupProbe.Bool())
-		require.Equal(t, expectReadinessProbe, statusAnn.States.ReadinessProbe.Bool())
-		require.Equal(t, podcommon.StateContainerRunning, statusAnn.States.Container)
-		require.Equal(t, podcommon.StateBoolTrue, statusAnn.States.Started)
-		require.Equal(t, podcommon.StateBoolTrue, statusAnn.States.Ready)
-		require.Equal(t, podcommon.StateResourcesPostStartup, statusAnn.States.Resources)
-		require.Equal(t, podcommon.StateStatusResourcesContainerResourcesMatch, statusAnn.States.StatusResources)
-
-		require.NotEmpty(t, statusAnn.Scale.LastCommanded)
-		require.NotEmpty(t, statusAnn.Scale.LastEnacted)
-		require.Empty(t, statusAnn.Scale.LastFailed)
-	}
-}
-
-func ensureEvents(t *testing.T, reason string, substrs []string, namespace string, names []string) {
-	for _, name := range names {
-		messages, err := kubeGetEventMessages(t, namespace, name, reason)
-		maybeLogErrAndFailNow(t, err)
-
-		for _, substr := range substrs {
-			gotMessage := false
-			for _, message := range messages {
-				if strings.Contains(message, substr) {
-					gotMessage = true
-					break
-				}
-			}
-
-			assert.True(t, gotMessage)
-		}
-	}
-}
-
 func maybeRegisterCleanup(t *testing.T, namespace string) {
 	if suppliedConfig.deleteNsPostTest {
 		t.Cleanup(func() {
 			_ = kubeDeleteNamespace(t, namespace)
 		})
-	}
-}
-
-func maybeLogErrAndFailNow(t *testing.T, err error) {
-	if err != nil {
-		logMessage(t, err)
-		t.FailNow()
 	}
 }
